Redo home page in tsx
